Move file i/o out of benchmarked code (#43)
Replaces the file read within each test with a simple memory copy. No
reason to include the time it takes to read the file in the benchmark.
This doesn't have a major impact, but it is just enough to be measurable
on my machine (difference of about 2 standard deviations on some of the
lex/parse/nu_old tests).