Add lexing of string interpolation (#44)
Adds lexing of string interpolations, both single- and double-quoted. It
required some manual parsing, which took a few tries, but it's not too
bad and can serve as a proof-of-concept if we need some more manual
lexing in the Logos lexer.

One limitation is that the string intepolations cannot host the same
types of quotes, for example `$"foo("bar")baz"` is not allowed, but in
Nushell it is allowed. It might be possible to add it, but would require
more hassle, so I left it out for now as it's not a severe limitation.

It also adds a LexError for reporting lexing errors. Now it is used only
in string interpolation, but it can be expanded in the future.