Add some operators (#45)
As part of https://github.com/nushell/new-nu-parser/issues/21, this PR
adds binary operators that are either symbols or single barewords. I
couldn't figure out an easy way to handle operators that span multiple
tokens, such as `starts-with`. I also didn't add `not`.

I threw in some code for `++=` but no tests. I assume we'll get to that
once we implement typechecking for assignment oeprators, which this PR
doesn't do.