Parse command return types (#46)
This PR parses return types for commands and records these in the
typechecker, although it doesn't check that the actual result type
matches the expected type.

Regarding parsing, I chose to add a `AstNode::InOutTypes(Vec<NodeId>)`
variant to hold the list of input-output type pairs. Making that list
its own node means we can highlight the entire span `[ foo -> bar ]` in
error messages, including the brackets. But if that's not something we
need, I can get rid of the `InOutTypes` variant and directly store a
`Vec<NodeId>` inside each `AstNode::Def`. That'd be simpler.